ProSampler: An ultra-fast motif finding tool on large ChIP-seq datasets
ProSampler is a motif finding program aiming at identifying motifs of the ChIP-ed TF as well as the cooperative TFs on large ChIP-seq datasets. It was developed by the research group of Zhengchang Su in UNCC and Guojun Li in SDU. Compared to the existing motif finding tools, it possesses high speed and accuracy with the capability of identifying motif lengths.